Move RLock mock to be explicitly private
