+++ src/urllib3/_private_compat/rlock.py
